Mostly worked on CSS
Added a color gradient under the links when they are hovered over in the shape of a bar. The I just added those styles into the transition styles with the width as 100% and the opacity to 1. Then I went back to the transition styles and added a transition that applies the opacity and width over half a second interval.